fixed soap include bug
git-svn-id: http://svn.ilias.de/svn/ilias/trunk@8667 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ webservice/soap/classes/class.ilSoapUserAdministration.php
@@ -91,7 +91,7 @@ class ilSoapUserAdministration extends ilSoapAdministration
 			return $this->__raiseError('Check access failed.'.USER_FOLDER_ID,'Server');
 		}
 
-		return (int) ilObjUser::getUserIdByLogin($user_name);
+		return ilObjUser::getUserIdByLogin($user_name);
 	}
 
 	function getUser($sid,$user_id)
